removed mbot.c/h
1. removed mbot.c/h
2. mbot_classic tested
3. next to test omni, once passed the driving test, then next move to param type
