+++ src/mbot_classic.h
@@ -1,5 +1,25 @@
-#include "mbot.h"
+#ifndef MBOT_CLASSIC_H
+#define MBOT_CLASSIC_H
+
+#include <pico/stdlib.h>
+#include <pico/mutex.h>
+#include <pico/time.h>
+#include <hardware/gpio.h>
+#include <mbot/motor/motor.h>
+#include <mbot/encoder/encoder.h>
+#include <mbot/motor/motor.h>
+#include <mbot/defs/mbot_pins.h>
 #include <mbot/defs/mbot_params.h>
+#include <mbot/fram/fram.h>
+#include <mbot/imu/imu.h>
+#include <mbot/utils/utils.h>
+#include <rc/math/filter.h>
+#include <rc/mpu/mpu.h>
+#include <math.h>
+#include <inttypes.h>
+#include "mbot_print.h"
+#include "mbot_comms.h"
+#include "mbot_odometry.h"
 
 // Hardware Parameters
 #define GEAR_RATIO              78.0
@@ -12,3 +32,5 @@
 #define DIFF_MOTOR_LEFT_SLOT         0    // Left motor using M0 slot
 #define DIFF_MOTOR_RIGHT_SLOT        1    // Right motor using M1 slot
 #define UNUSED_DIFF_MOTOR_SLOT       2    // defined for mbot classic, 2 means M2 slot
+
+#endif
