+++ src/mbot_omni.c
@@ -2,6 +2,7 @@
  * This file is the main executable for the MBot OMNI firmware.
  */
 #include "mbot_omni.h"
+#pragma pack(1)
 
 // Global Variables and Forward Declarations
 static bool running = false;
@@ -12,9 +13,16 @@ uint64_t global_pico_time = 0;
 bool global_comms_status = COMMS_ERROR;
 int drive_mode = 0;
 mbot_bhy_config_t mbot_imu_config;
+mbot_bhy_data_t mbot_imu_data;
 
 // Forward declaration for internal helper function
+void mbot_read_imu(serial_mbot_imu_t *imu);
+int mbot_init_hardware(void);
 static float _calibrated_pwm_from_vel_cmd(float vel_cmd, int motor_idx);
+void print_mbot_params(const mbot_params_t* params);
+void mbot_calculate_motor_vel(serial_mbot_encoders_t encoders, serial_mbot_motor_vel_t *motor_vel);
+int mbot_init_pico(void);
+void mbot_read_encoders(serial_mbot_encoders_t* encoders);
 
 /*********************************************************************
  * Main Control Functions
@@ -152,6 +160,60 @@ int main()
  * They are not intended for modification by students. These functions
  * provide lower-level control and utility support.
  ******************************************************/
+void mbot_read_encoders(serial_mbot_encoders_t* encoders){
+    int64_t delta_time = global_utime - encoders->utime;
+    encoders->utime = global_utime;
+    encoders->delta_time = delta_time;
+
+    encoders->ticks[params.mot_right] = mbot_encoder_read_count(params.mot_right);
+    encoders->delta_ticks[params.mot_right] = mbot_encoder_read_delta(params.mot_right);
+    encoders->ticks[params.mot_left] = mbot_encoder_read_count(params.mot_left);
+    encoders->delta_ticks[params.mot_left] = mbot_encoder_read_delta(params.mot_left);
+    encoders->ticks[params.mot_back] = mbot_encoder_read_count(params.mot_back);
+    encoders->delta_ticks[params.mot_back] = mbot_encoder_read_delta(params.mot_back);
+}
+
+int mbot_init_pico(void){
+    bi_decl(bi_program_description("Firmware for the MBot Robot Control Board"));
+    
+    // set master clock to 250MHz (if unstable set SYS_CLOCK to 125Mhz)
+    if(!set_sys_clock_khz(125000, true)){
+        printf("ERROR mbot_init_pico: cannot set system clock\n");
+        return MBOT_ERROR;
+    }; 
+    
+    stdio_init_all(); // enable USB serial terminal
+    sleep_ms(500);
+    printf("\nMBot Booting Up!\n");
+    return MBOT_OK;
+}
+
+void mbot_calculate_motor_vel(serial_mbot_encoders_t encoders, serial_mbot_motor_vel_t *motor_vel){
+    float conversion = (1.0 / params.gear_ratio) * (1.0 / params.encoder_resolution) * 1E6f * 2.0 * M_PI;
+    motor_vel->velocity[params.mot_left] = params.encoder_polarity[params.mot_left] * (conversion / encoders.delta_time) * encoders.delta_ticks[params.mot_left];
+    motor_vel->velocity[params.mot_back] = params.encoder_polarity[params.mot_back] * (conversion / encoders.delta_time) * encoders.delta_ticks[params.mot_back];
+    motor_vel->velocity[params.mot_right] = params.encoder_polarity[params.mot_right] * (conversion / encoders.delta_time) * encoders.delta_ticks[params.mot_right];
+}
+
+void mbot_read_imu(serial_mbot_imu_t *imu){
+    imu->utime = global_utime;
+    imu->gyro[0] = mbot_imu_data.gyro[0];
+    imu->gyro[1] = mbot_imu_data.gyro[1];
+    imu->gyro[2] = mbot_imu_data.gyro[2];
+    imu->accel[0] = mbot_imu_data.accel[0];
+    imu->accel[1] = mbot_imu_data.accel[1];
+    imu->accel[2] = mbot_imu_data.accel[2];
+    imu->mag[0] = mbot_imu_data.mag[0];
+    imu->mag[1] = mbot_imu_data.mag[1];
+    imu->mag[2] = mbot_imu_data.mag[2];
+    imu->angles_rpy[0] = mbot_imu_data.rpy[0];
+    imu->angles_rpy[1] = mbot_imu_data.rpy[1];
+    imu->angles_rpy[2] = mbot_imu_data.rpy[2];
+    imu->angles_quat[0] = mbot_imu_data.quat[0];
+    imu->angles_quat[1] = mbot_imu_data.quat[1];
+    imu->angles_quat[2] = mbot_imu_data.quat[2];
+    imu->angles_quat[3] = mbot_imu_data.quat[3];   
+}
 
 int mbot_init_hardware(void){
     sleep_ms(1000);
@@ -177,8 +239,8 @@ int mbot_init_hardware(void){
     return MBOT_OK;
 }
 
-// Use slope + intercept from calibration to generate a PWM command.
 float _calibrated_pwm_from_vel_cmd(float vel_cmd, int motor_idx){
+    // Use slope + intercept from calibration to generate a PWM command.
     if (vel_cmd > 0.0)
     {
         return (vel_cmd * params.slope_pos[motor_idx]) + params.itrcpt_pos[motor_idx];
@@ -188,4 +250,21 @@ float _calibrated_pwm_from_vel_cmd(float vel_cmd, int motor_idx){
         return (vel_cmd * params.slope_neg[motor_idx]) + params.itrcpt_neg[motor_idx];
     }
     return 0.0;
+}
+
+void print_mbot_params(const mbot_params_t* params) {
+    printf("Robot Type: %d\n", params->robot_type);
+    printf("Wheel Radius: %f\n", params->wheel_radius);
+    printf("Wheel Base Radius: %f\n", params->wheel_base_radius);
+    printf("Gear Ratio: %f\n", params->gear_ratio);
+    printf("Encoder Resolution: %f\n", params->encoder_resolution);
+    printf("Motor Left: %d\n", params->mot_left);
+    printf("Motor Right: %d\n", params->mot_right);
+    printf("Motor Back: %d\n", params->mot_back);
+    printf("Motor Polarity: %d %d %d\n", params->motor_polarity[0], params->motor_polarity[1], params->motor_polarity[2]);
+    printf("Encoder Polarity: %d %d %d\n", params->encoder_polarity[0], params->encoder_polarity[1], params->encoder_polarity[2]);
+    printf("Positive Slope: %f %f %f\n", params->slope_pos[0], params->slope_pos[1], params->slope_pos[2]);
+    printf("Positive Intercept: %f %f %f\n", params->itrcpt_pos[0], params->itrcpt_pos[1], params->itrcpt_pos[2]);
+    printf("Negative Slope: %f %f %f\n", params->slope_neg[0], params->slope_neg[1], params->slope_neg[2]);
+    printf("Negative Intercept: %f %f %f\n", params->itrcpt_neg[0], params->itrcpt_neg[1], params->itrcpt_neg[2]);
 }
